projects/make-recall-decisions-and-delius/src/main/kotlin/uk/gov/justice/digital/hmpps/service/ProviderService.kt
package uk.gov.justice.digital.hmpps.service

import org.springframework.stereotype.Service
import uk.gov.justice.digital.hmpps.api.model.toProvider
import uk.gov.justice.digital.hmpps.integrations.delius.provider.entity.ProviderRepository

@Service
class ProviderService(private val providerRepository: ProviderRepository) {
    fun getProviders() = providerRepository.findActive().map { it.toProvider() }
    fun getProviderByCode(code: String) = providerRepository.findByCode(code)?.toProvider()
}
